requirements.txt
APScheduler==3.6.3
black==23.9.1
cachetools==4.2.2
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-commas==2.1.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pep8-naming==0.13.3
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
six==1.16.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
